pyproject.toml
[project]
name = "yolovex"
version = "0.1.0"
description = "Interactive YOLO architecture explorer"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pyqt5>=5.15.11",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "typer>=0.20.0",
    "ultralytics>=8.4.42",
]

[project.scripts]
yolovex = "yolovex.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yolovex"]
